Unreviewed, fix debug build warning.
%llu needs 'long long unsigned int'.

Type casting unit64_t to long long unsigned int.

* Modules/indexeddb/client/IDBOpenDBRequestImpl.cpp:
(WebCore::IDBClient::IDBOpenDBRequest::onUpgradeNeeded):
* Modules/indexeddb/server/UniqueIDBDatabase.cpp:
(WebCore::IDBServer::UniqueIDBDatabase::notifyConnectionsOfVersionChange):

Canonical link: https://commits.webkit.org/168213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190926 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/indexeddb/client/IDBOpenDBRequestImpl.cpp
@@ -84,7 +84,7 @@ void IDBOpenDBRequest::onUpgradeNeeded(const IDBResultData& resultData)
     uint64_t oldVersion = database->info().version();
     uint64_t newVersion = transaction->info().newVersion();
 
-    LOG(IndexedDB, "IDBOpenDBRequest::onUpgradeNeeded() - current version is %llu, new is %llu", oldVersion, newVersion);
+    LOG(IndexedDB, "IDBOpenDBRequest::onUpgradeNeeded() - current version is %llu, new is %llu", static_cast<unsigned long long>(oldVersion), static_cast<unsigned long long>(newVersion));
 
     m_result = IDBAny::create(WTF::move(database));
     m_readyState = IDBRequestReadyState::Done;
